requirements.txt
google.generativeai
python-dotenv
flask
flask_cors
ollama
groq
sounddevice
numpy
anthropic
openai
git+https://github.com/openai/whisper.git
silero_vad
requests
soundfile
Pillow
faster-whisper
language_tool_python
PyGObject
coqui-tts
aiohttp